gestao-reservas-projeto-completo/gestao-reservas-backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-cors-headers>=4.4
PyJWT>=2.8

dj-database-url>=2.1
psycopg2-binary>=2.9

black>=24.0
flake8>=7.0

pre-commit>=3.7
